use build tag rocksdb_v6
